requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.25.0
config==0.3.9
confusable-homoglyphs==3.2.0
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-v5==1.0.3
django-bootstrap3==15.0.0
django-config==0.1.4
django-csv==1.2
django-decouple==2.1
django-gunicorn==0.1.1
django-heroku==0.3.1
django-registration==2.4.1
django-shell==1.0.2
django-tinymce==3.3.0
django-utils-six==2.0
gunicorn==19.9.0
importlib-metadata==2.1.1
Pillow==8.2.0
psycopg2==2.8.6
python-config==0.1.2
python-decouple==3.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
zipp==3.4.1
